#48 파이어베이스 관리를 위한 클래스 생성
