+++ attendance-ios/Source/SignUp/SignUpViewModel.swift
@@ -41,6 +41,7 @@ final class SignUpViewModel: ViewModel {
     let disposeBag = DisposeBag()
 
     private let configWorker = ConfigWorker()
+    private let firebaseWorker = FirebaseWorker()
 
     init() {
         setupConfig()
@@ -105,33 +106,16 @@ extension SignUpViewModel {
               let teamType = try? input.teamType.value(),
               let teamNumber = try? input.teamNumber.value() else { return }
 
-        let db = Firestore.firestore()
-        let docRef = db.collection("member")
+        let newUser = FirebaseNewUser(name: name, positionType: positionType, teamType: teamType, teamNumber: teamNumber)
 
-        UserApi.shared.me { [weak self] user, error in
-            guard let self = self, let user = user, let userId = user.id else { return }
-
-            docRef.document("\(userId)").setData([
-                "id": userId,
-                "name": name,
-                "position": positionType.rawValue,
-                "team": ["number": teamNumber, "type": teamType.upperCase],
-                "attendances": self.makeEmptyAttendances()
-            ]) { [weak self] error in
-                guard error == nil else { return }
-                self?.output.goToHome.accept(())
+        func registerInfo() {
+            firebaseWorker.registerInfo(newUser: newUser) { [weak self] result in
+                switch result {
+                case .success: self?.output.goToHome.accept(())
+                case .failure: ()
+                }
             }
         }
-    }
-
-    private func makeEmptyAttendances() -> [[String: Any]] {
-        var attendances: [[String: Any]] = []
-        let sessionCount = 20
-        for id in 0..<sessionCount {
-            let empty: [String: Any] = ["sessionId": id, "attendanceType": ["text": "결석", "point": -20]]
-            attendances.append(empty)
-        }
-        return attendances
     }
 
 }
